feat: Metadata: 自定义指标查询路由过期时间根据关联集群获取 #1010158081128171911

+++ pkg/bk-monitor-worker/internal/metadata/service/spaceredis.go
@@ -1505,7 +1505,7 @@ func (s *SpacePusher) filterTsInfo(bkTenantId string, tableIds []string) (*TsInf
 		if !ok {
 			continue
 		}
-		vmClusterRetentionTimeMap[cluster.ClusterID] = retentionTime.(int)
+		vmClusterRetentionTimeMap[cluster.ClusterID] = int(retentionTime.(float64))
 	}
 
 	// 获取tableId与VM集群的对应关系
